+++ concordium-sdk-examples/src/main/java/com/concordium/sdk/examples/contractexample/cis2nft/Cis2Nft.java
@@ -70,7 +70,7 @@ public class Cis2Nft implements Callable<Integer> {
                 .timeout(timeout)
                 .build();
         var client = ClientV2.from(connection);
-        Nonce nonce = client.getAccountInfo(BlockQuery.BEST, AccountQuery.from(AccountAddress.from(SENDER_ADDRESS))).getAccountNonce();
+        Nonce nonce = client.getAccountInfo(BlockQuery.BEST, AccountQuery.from(AccountAddress.from(SENDER_ADDRESS))).getNonce();
 
         switch (this.methodName) {
             case INIT:
@@ -119,7 +119,7 @@ public class Cis2Nft implements Callable<Integer> {
                 .sender(AccountAddress.from(SENDER_ADDRESS))
                 .payload(payload)
                 .expiry(EXPIRY)
-                .nonce(AccountNonce.from(nonce))
+                .nonce(nonce)
                 .signer(SIGNER)
                 .maxEnergyCost(UInt64.from(10000))
                 .build();
@@ -135,7 +135,7 @@ public class Cis2Nft implements Callable<Integer> {
                 .sender(AccountAddress.from(SENDER_ADDRESS))
                 .payload(payload)
                 .expiry(EXPIRY)
-                .nonce(AccountNonce.from(nonce))
+                .nonce(nonce)
                 .signer(SIGNER)
                 .maxEnergyCost(UInt64.from(10000))
                 .build();