+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/ConfigureBakerPayload.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.transactions;
 
+import com.concordium.sdk.responses.transactionstatus.PartsPerHundredThousand;
 import com.concordium.sdk.types.UInt16;
 import com.concordium.sdk.types.UInt32;
 import lombok.Builder;
@@ -8,6 +9,7 @@ import lombok.Getter;
 import lombok.val;
 
 import java.nio.ByteBuffer;
+import java.util.Objects;
 
 /**
  * Payload for configuring a baker.
@@ -30,7 +32,7 @@ public class ConfigureBakerPayload {
     /**
      * Whether the pool is open for delegators.
      */
-    private final Integer openForDelegation;
+    private final Boolean openForDelegation;
     /**
      * The key/proof pairs to verify the baker.
      */
@@ -39,18 +41,22 @@ public class ConfigureBakerPayload {
      * The URL referencing the baker's metadata.
      */
     private final String metadataUrl;
+
     /**
      * The commission the pool owner takes on transaction fees.
+     * The supplied value is interpreted as "the value" / 100_000
      */
-    private final UInt32 transactionFeeCommission;
+    private final PartsPerHundredThousand transactionFeeCommission;
     /**
      * The commission the pool owner takes on baking rewards.
+     * The supplied value is interpreted as "the value" / 100_000
      */
-    private final UInt32 bakingRewardCommission;
+    private final PartsPerHundredThousand bakingRewardCommission;
     /**
      * The commission the pool owner takes on finalization rewards.
+     * The supplied value is interpreted as "the value" / 100_000
      */
-    private final UInt32 finalizationRewardCommission;
+    private final PartsPerHundredThousand finalizationRewardCommission;
 
     ByteBuffer createNotNullBuffer(byte[] bufferBytes) {
         val buffer = ByteBuffer.allocate(bufferBytes.length);
@@ -64,19 +70,19 @@ public class ConfigureBakerPayload {
 
         bitValue |= ((this.capital != null) ? it : 0);
         it *= 2;
-        bitValue |= ((this.restakeEarnings != null) ? it : 0);
+        bitValue |= (!(Objects.isNull(this.restakeEarnings)) ? it : 0);
         it *= 2;
-        bitValue |= ((this.openForDelegation != null) ? it : 0);
+        bitValue |= (!Objects.isNull(this.openForDelegation) ? it : 0);
         it *= 2;
         bitValue |= ((this.keysWithProofs != null) ? it : 0);
         it *= 2;
         bitValue |= ((this.metadataUrl != null) ? it : 0);
         it *= 2;
-        bitValue |= ((this.transactionFeeCommission != null) ? it : 0);
+        bitValue |= (!Objects.isNull(this.transactionFeeCommission ) ? it : 0);
         it *= 2;
-        bitValue |= ((this.bakingRewardCommission != null) ? it : 0);
+        bitValue |= (!Objects.isNull(this.bakingRewardCommission) ? it : 0);
         it *= 2;
-        bitValue |= ((this.finalizationRewardCommission != null) ? it : 0);
+        bitValue |= (!Objects.isNull(this.finalizationRewardCommission) ? it : 0);
 
         return UInt16.from(bitValue).getBytes();
     }
@@ -109,7 +115,7 @@ public class ConfigureBakerPayload {
         }
 
         if (this.openForDelegation != null) {
-            val openForDelegationByte = this.openForDelegation.byteValue();
+            val openForDelegationByte = (byte) (this.openForDelegation ? 1 : 0);
             openForDelegationBuffer = createNotNullBuffer(new byte[]{openForDelegationByte});
             bufferLength += TransactionType.BYTES;
         }