+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/ConfigureDelegationTransaction.java
@@ -2,6 +2,7 @@ package com.concordium.sdk.transactions;
 
 import com.concordium.sdk.exceptions.TransactionCreationException;
 import com.concordium.sdk.types.AccountAddress;
+import com.concordium.sdk.types.Nonce;
 import lombok.*;
 
 /**
@@ -15,7 +16,7 @@ public class ConfigureDelegationTransaction extends AccountTransaction {
     private ConfigureDelegationTransaction(
             @NonNull final ConfigureDelegationPayload payload,
             @NonNull final AccountAddress sender,
-            @NonNull final AccountNonce nonce,
+            @NonNull final Nonce nonce,
             @NonNull final Expiry expiry,
             @NonNull final TransactionSigner signer) {
         super(sender, nonce, expiry, signer, ConfigureDelegation.createNew(payload), TransactionTypeCost.CONFIGURE_DELEGATION.getValue());
@@ -36,7 +37,7 @@ public class ConfigureDelegationTransaction extends AccountTransaction {
     @Builder
     public static ConfigureDelegationTransaction from(final ConfigureDelegationPayload payload,
                                                       final AccountAddress sender,
-                                                      final AccountNonce nonce,
+                                                      final Nonce nonce,
                                                       final Expiry expiry,
                                                       final TransactionSigner signer) {
         try {