+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransferToEncryptedTransaction.java
@@ -3,6 +3,7 @@ package com.concordium.sdk.transactions;
 
 import com.concordium.sdk.exceptions.TransactionCreationException;
 import com.concordium.sdk.types.AccountAddress;
+import com.concordium.sdk.types.Nonce;
 import lombok.*;
 
 /**
@@ -15,7 +16,7 @@ public class TransferToEncryptedTransaction extends AccountTransaction {
     private TransferToEncryptedTransaction(
             @NonNull final CCDAmount amount,
             @NonNull final AccountAddress sender,
-            @NonNull final AccountNonce nonce,
+            @NonNull final Nonce nonce,
             @NonNull final Expiry expiry,
             @NonNull final TransactionSigner signer) {
         super(sender, nonce, expiry, signer, TransferToEncrypted.createNew(amount), TransactionTypeCost.TRANSFER_TO_ENCRYPTED.getValue());
@@ -37,7 +38,7 @@ public class TransferToEncryptedTransaction extends AccountTransaction {
     public static TransferToEncryptedTransaction from(
             final CCDAmount amount,
             final AccountAddress sender,
-            final AccountNonce nonce,
+            final Nonce nonce,
             final Expiry expiry,
             final TransactionSigner signer) {
         try {