+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 # Unreleased changes
+ - Cleanup the API a bit for configure baker transaction by using `PartsPerHundredThousands` for determining the commission rates.
+ - Fix serialization of `ConfigureDelegation` transaction
+ - Remove `AccountNonce` in favor of just using the `Nonce` type across the API.
  - Fix a bug in the serialization of `AccountIndex`
  - Fix a bug that caused `getAccountInfo` to fail for delegator and baker accounts if they had no stake pending changes. 
    This change is also propagated to the type level such that `Baker` and `AccountDelegation` retains an `Optional<PendingChange>`